# contract-service-rest/src/main/resources/swagger/contract-service-rest-swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Simple Contract Calculation Application",
    "version": "1.0.0",
    "title": "Swagger Contract Calculator"
  },
  "host": "localhost",
  "basePath": "/v1",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/contract": {
      "post": {
        "tags": [
          "contract"
        ],
        "summary": "calculate contract",
        "description": "",
        "operationId": "calulateContract",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Contract to be calculated",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Contract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "contract sucessfully calculated",
            "schema": {
              "$ref": "#/definitions/Contract"
            }
          },
          "405": {
            "description": "invalid contract input"
          },
          "500": {
            "description": "server error"
          }
        }
      }
    }
  },
  "definitions": {
    "Contract": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "readOnly":true
        },
        "type": {
          "type": "string"
        },
        "premium": {
          "type": "number",
          "format": "double",
          "readOnly":true
        },
        "signed": {
          "type": "string",
          "format": "date"
        },
        "clients": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Client"
          },
          "minItems": 1
        },
        "risk": {
          "type": "string",
          "description": "Order Status",
          "enum": [
            "low",
            "medium",
            "high"
          ],
          "readOnly":true
        },
        "active": {
          "type": "boolean",
          "default": false,
          "readOnly":true
        }
      },
      "required": [
        "signed"
      ]
    },
    "Client": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "$ref": "#/definitions/Name"
        },
        "dateOfBirth": {
          "type": "string",
          "format": "date"
        }
      },
      "required": [
        "id",
        "dateOfBirth"
      ]
    },
    "Name": {
      "type": "object",
      "properties": {
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        }
      }
    }
  }
}